Seperated into new module also added error page
