+++ .gitignore
@@ -1,2 +1,1 @@
 /node_modules
-config.json
