+++ config.json
@@ -1,0 +1,10 @@
+{
+	"request_token_url": "https://api.twitter.com/oauth/request_token",
+	"access_token_url": "https://api.twitter.com/oauth/access_token",
+	"authorize_url": "https://twitter.com/oauth/authorize",
+	"consumer_key": "xxxxxxxxxxxxxxxxxxxxxxxxxxxx",
+	"consumer_secret": "xxxxxxxxxxxxxxxxxxxxxxxxx",
+	"oauth_version": "1.0",
+	"oauth_signature": "HMAC-SHA1",
+	"oauth_callback": "http://localhost:3000/request-token"
+}
