+++ server.js
@@ -7,7 +7,7 @@ var cookieParser = require('cookie-parser');
 var path = require('path');
 var config = require('./config');
 var bodyParser = require('body-parser');
-
+var twitter = require('./twitter');
 var url = require('url');
 
 // Create the oauth object for accessing Twitter
@@ -42,112 +42,9 @@ app.get('/', function(req, res) {
 	if (!req.cookies.access_token || !req.cookies.access_token_secret || !req.cookies.twitter_id) {
 		res.render('login');
 	} else {
-		getTwitterData(req, res);
+		twitter.getTwitterData(req, res, oauth);
 	} // If 
 }); // oauth.get
-
-function getTwitterData(req, res) {
-
-	// Access cookie details	
-	var access_token = req.cookies.access_token;
-	var access_secret = req.cookies.access_token_secret;
-	var name  = req.cookies.twitter_id;
-
-	var followers = [];
-	var tweets = [];
-	var dmessages  = [];
-
-	var userTimeLine = new Promise(function(resolve, reject) {
-		// Call twitter RESt API to get user timeline
-		var url = 'https://api.twitter.com/1.1/statuses/user_timeline.json?count=5';
-		oauth.get(url, access_token, access_secret, function(error, data) {
-
-			if(!error) {
-				// Parse JSON response
-				data = JSON.parse(data);
-				tweets = [];
-
-				// Iterate through array and store into tweets array
-				data.forEach(function(value, index) {
-					// Temporary tweet object to store data
-					var obj = {};
-
-					obj.text = value.text;
-					obj.name = value.user.name;
-					obj.sname = value.user.screen_name;
-					obj.image = value.user.profile_image_url;
-					obj.rtcount = value.retweet_count;
-					obj.favcount = value.favorite_count;
-
-					tweets.push(obj);
-				}); // End of forEach
-
-				resolve(tweets);
-			} else {
-				reject(error);
-			}
-		}); // End of oauth.get - user timeline
-	}); // End of userTimeLine
-
-	var userFriends = new Promise(function(resolve, reject) {
-
-		// Call twitter RESt API to get user friends
-		url = 'https://api.twitter.com/1.1/friends/list.json?count=5';
-		oauth.get(url, access_token, access_secret, function(error, data) {
-
-			if (!error) {
-				data = JSON.parse(data);
-
-				followers = [];
-				data.users.forEach(function(value) {
-					var obj = {};
-					obj.name = value.name;
-					obj.sname = value.screen_name;
-					obj.image = value.profile_image_url;
-					obj.following = value.following;
-					followers.push(obj);
-				}); // End of forEach
-
-				resolve(followers);
-			} else {
-				reject(error);
-			}
-		}); // End of oauth.get - user friends
-	}); // End of userFriends
-
-	var userDMessage = new Promise(function(resolve, reject) {
-
-		// Call twitter RESt API to get user Direct Message 
-		// url = 'https://api.twitter.com/1.1/direct_messages/sent.json?count=5';
-		url = 'https://api.twitter.com/1.1/direct_messages.json?count=5';
-		oauth.get(url, access_token, access_secret, function(error, data) {
-
-			if (!error) {
-				data = JSON.parse(data);
-
-				dmessages = [];
-				data.forEach(function(value) {
-					var obj = {};
-					obj.name = value.sender_screen_name;
-					obj.text = value.text;
-					obj.image = value.sender.profile_image_url;
-					dmessages.push(obj);
-				});		
-				resolve(dmessages);
-			} else {
-				reject(error);
-			}
-		}); // End of oauth.get - user Direct Message
-	}); // End of userDMessage
-
-	Promise.all([userTimeLine, userFriends, userDMessage]).then(function(data) {
-		res.render('index', 
-			{name: data[0][0].sname, tweets:data[0], followers:data[1], dmessages:data[2]});
-	})
-	.catch(function(error) {
-		res.send('something wrong with getTweetData');
-	})
-} // End of getTwitterData
 
 // 2. When user Click to Login - This route will receive the request
 app.get('/auth/twitter', function(req, res) {
@@ -216,6 +113,7 @@ app.post('/post', function(req, res) {
         	res.redirect('/');
         } else {
         	console.log('Error: Something is wrong.\n'+JSON.stringify(error)+'\n');
+        	res.render('error', {error: error});
         }
     }); // ouath.post
 }); // app.post
