+++ views/error.jade
@@ -1,0 +1,8 @@
+doctype html
+html(lang="en")
+  head
+    link(rel='stylesheet', href='css/main.css')
+  body
+    .error
+      h3 #{error}
+
