+++ views/layout.jade
@@ -5,7 +5,7 @@ head
       ––––––––––––––––––––––––––––––––––––––––––––––––––
     meta(charset='utf-8')
     |   
-    title Twitter Client
+    title Tweet App
     |   
     meta(name='description', content='')
     |
