+++ views/login.jade
@@ -1,7 +1,8 @@
 doctype html
 html(lang="en")
   head
-    link(rel='stylesheet', href='css/login.css')
+    title Tweet App
+    link(rel='stylesheet', href='css/main.css')
   body
     div
       header
